requirements.txt
flask==3.1.3
supabase==2.29.0
python-dotenv==1.2.2
waitress==3.0.2
